+++ changelog.txt
@@ -24,10 +24,16 @@ version 1.00
 После внесения события предлагается команда /showdata, которая показывает последнее внесенное событие.
 
 Просмотр событий:
-- выдает список событий
+- выдает список событий отсортированных по порядке внесения
 - PDF с полным описанием всех параметров событий
 Данные отдаются по user_id, только временные события.
 
 Инфраструктурно:
 - настроено логирование через logging
 - для временного хранения используется MemoryStorage() aiogram
+
+version 1.01
+
+- docker-compose restart=always в контейнерах postgres и pgadmin
+- добавлена кнопка "Назад" в меню внесения событий
+- ограничение выдачи Временных событий (temporary) по текущей дате
